feat: add viewmodel for shops fragment

+++ app/src/main/java/com/example/secretcustomer/data/Shop.kt
@@ -12,5 +12,5 @@ data class Shop(
     @SerializedName("address")
     val address: String,
     @SerializedName("balance")
-    val balance: Float
+    val balance: Int
 )
